Move ImmediateWorker to its own module
