+++ bot/multithreading/worker/__init__.py
@@ -43,17 +43,3 @@ class AbstractWorker(Worker):
 
     def shutdown(self):
         raise NotImplementedError()
-
-
-class ImmediateWorker(AbstractWorker):
-    def __init__(self, error_handler: callable):
-        super().__init__("immediate", error_handler)
-
-    def run(self):
-        pass
-
-    def post(self, work: Work):
-        self._work(work)
-
-    def shutdown(self):
-        pass
